Moves the config to a directory
